requirements.txt
discord.py==2.4.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.1
